(PC-34760) feat(cheatcodes): review and move queries
