+++ doc/decision-records/DR003 - queries.md
@@ -3,3 +3,12 @@
 ## Audit Folders
 
 ### Cheatcodes
+
+- useSomeOfferIdQuery
+- useFeatureFlagAllQuery
+- useVenue => CheatcodesScreenAccesLibre
+
+## App State
+
+- useDeviceInfo => dans un hook pas stocké sur un Store
+- storage !! => stockage persisté de l'app
